+++ tests/tinytest/test_bind_cols_lazy.R
@@ -1,12 +1,16 @@
+### [GENERATED AUTOMATICALLY] Update test_bind_cols.R instead.
+
+Sys.setenv('TIDYPOLARS_TEST' = TRUE)
+
 source("helpers.R")
 using("tidypolars")
 
 l <- list(
-  polars::pl$DataFrame(
+  polars::pl$LazyFrame(
     x = sample(letters, 20),
     y = sample(1:100, 20)
   ),
-  polars::pl$DataFrame(
+  polars::pl$LazyFrame(
     a = sample(letters, 20),
     z = sample(1:100, 20)
   )
@@ -18,16 +22,16 @@ expect_dim(
 )
 
 
-p1 <- pl$DataFrame(
+p1 <- pl$LazyFrame(
   x = sample(letters, 20),
   y = sample(1:100, 20)
 )
-p2 <- pl$DataFrame(
+p2 <- pl$LazyFrame(
   z = sample(letters, 20),
   w = sample(1:100, 20)
 )
 
-expect_equal(
+expect_equal_lazy(
   bind_cols_polars(p1, p2),
   bind_cols_polars(list(p1, p2))
 )
@@ -37,13 +41,13 @@ l3 <- list(
     x = sample(letters, 20),
     y = sample(1:100, 20)
   ),
-  polars::pl$DataFrame(
+  polars::pl$LazyFrame(
     y = sample(letters, 20),
     z = sample(1:100, 20)
   )
 )
 
-expect_error(
+expect_error_lazy(
   bind_cols_polars(l3),
   "must be either"
 )
@@ -53,14 +57,14 @@ expect_error(
 
 l4 <- append(
   l,
-  polars::pl$DataFrame(
+  polars::pl$LazyFrame(
     v = sample(letters, 20),
     w = sample(1:100, 20)
   )
 )
 
 if (Sys.getenv("TIDYPOLARS_TEST") == "TRUE") {
-  expect_error(
+  expect_error_lazy(
     bind_cols_polars(l4),
     "doesn't work with more than two"
   )
@@ -77,50 +81,52 @@ if (Sys.getenv("TIDYPOLARS_TEST") == "TRUE") {
 # arg ".name_repair"
 
 l5 <- list(
-  pl$DataFrame(a = 1, x = 2, y = 3),
-  pl$DataFrame(z = 1, x = 2, y = 3)
+  pl$LazyFrame(a = 1, x = 2, y = 3),
+  pl$LazyFrame(z = 1, x = 2, y = 3)
 )
 
 options(rlib_message_verbosity = "quiet")
 
-expect_equal(
+expect_equal_lazy(
   names(bind_cols_polars(l5)),
   c("a", "x...2", "y...3", "z", "x...5", "y...6")
 )
 
-expect_equal(
+expect_equal_lazy(
   names(bind_cols_polars(l5, .name_repair = "universal")),
   c("a", "x...2", "y...3", "z", "x...5", "y...6")
 )
 
-expect_error(
+expect_error_lazy(
   bind_cols_polars(l5, .name_repair = "check_unique"),
   "Names must be unique"
 )
 
-expect_error(
+expect_error_lazy(
   bind_cols_polars(l5, .name_repair = "minimal"),
   "Either provide unique names or use"
 )
 
-expect_error(
+expect_error_lazy(
   bind_cols_polars(l5, .name_repair = "blahblah"),
   "must be one of"
 )
 
 l6 <- list(
-  pl$DataFrame(x = 1)$rename(list(" " = "x")),
-  pl$DataFrame(x = 1)$rename(list(" " = "x"))
+  pl$LazyFrame(x = 1)$rename(list(" " = "x")),
+  pl$LazyFrame(x = 1)$rename(list(" " = "x"))
 )
 
-expect_equal(
+expect_equal_lazy(
   names(bind_cols_polars(l6)),
   c(" ...1", " ...2")
 )
 
-expect_equal(
+expect_equal_lazy(
   names(bind_cols_polars(l6, .name_repair = "universal")),
   c("....1", "....2")
 )
 
 options(rlib_message_verbosity = "default")
+
+Sys.setenv('TIDYPOLARS_TEST' = FALSE)