+++ tests/tinytest/test_funs_string.R
@@ -230,6 +230,20 @@ expect_equal(
     pull(foo)
 )
 
+expect_equal(
+  mutate(test, foo = str_replace_all(x1, c("LL" = "ll", " " = "_"))) |>
+    pull(foo),
+  mutate(test_df, foo = str_replace_all(x1, c("LL" = "ll", " " = "_"))) |>
+    pull(foo)
+)
+
+expect_equal(
+  mutate(test, foo = str_replace_all(x1, c("LL" = "ll", "( )" = "\\1\\1"))) |>
+    pull(foo),
+  mutate(test_df, foo = str_replace_all(x1, c("LL" = "ll", "( )" = "\\1\\1"))) |>
+    pull(foo)
+)
+
 # TODO: https://github.com/pola-rs/polars/issues/12110
 # expect_equal(
 #   mutate(test, foo = str_replace_all(x1, "[aeiou]", toupper)) |>