+++ NEWS.md
@@ -19,10 +19,35 @@
 
 * Add support for arguments `names_prefix` and `names_sep` in `pivot_wider()`.
 
+* Add support for `tidyr::uncount()`.
+
+* All `*_join()` functions now work when `by` is a specification created by 
+  `dplyr::join_by()`. Notice that this is limited to equality joins for now.
+  
+* You can now use the "embrace" operator `{{ }}` to pass unquoted column names
+  (among other things) as arguments of custom functions. See the ["Programming
+  with dplyr" vignette](https://dplyr.tidyverse.org/dev/articles/programming.html)
+  for some examples.
+  
+* `bind_cols_polars()` now works with two `LazyFrame`s, but not more.
+
+* Add support for argument `.name_repair` in `bind_cols_polars()` (#74).
+
+* Support for `.env$` and `.data$` pronouns in expressions of `filter()`, 
+  `mutate()` and `summarize()`.
+  
+* Support named vector in the argument `pattern` of `str_replace_all()`, where
+  names are patterns and values are replacements.
+
 **Bug fixes**
 
 * `summarize()` no longer errors when `across(everything(), ...)` is used with
   `.by`.
+
+* All `*_join()` functions no longer error when a named vector is provided in 
+  the argument `by`.
+  
+* Expressions with values only are not named "literal" anymore.
 
 **Misc**
 
@@ -77,7 +102,7 @@
 
 * Faster tidy selection (#61).
 
-# tidypolars (0.3.0)
+# tidypolars 0.3.0
 
 `tidypolars` requires `polars` >= 0.10.0.